requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
python-docx==1.1.0
PyPDF2==3.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.3.7
python-magic==0.4.27
aiofiles==23.2.1
jinja2==3.1.2
regex==2023.10.3
nltk==3.8.1
textstat==0.7.3 